[REFACTOR-ORGANIZE] Reordenning classes
