--- app/src/main/java/com/example/liverpooltest/model/CustomUrlParam.kt
@@ -1,3 +1,0 @@
-package com.example.liverpooltest.model
-
-class CustomUrlParam
