+++ app/src/main/java/com/example/liverpooltest/model/data/MainModel.kt
@@ -1,6 +1,6 @@
 package com.example.liverpooltest.model.data
 
-import com.example.liverpooltest.model.Status
+import com.example.liverpooltest.model.others.Status
 
 data class MainModel(
     val pageType: String,
