+++ app/src/main/java/com/example/liverpooltest/model/others/DwPromotionInfo.kt
@@ -1,4 +1,4 @@
-package com.example.liverpooltest.model
+package com.example.liverpooltest.model.others
 
 data class DwPromotionInfo(
     val dWPromoDescription: String,
