+++ app/src/main/java/com/example/liverpooltest/model/others/MetaData.kt
@@ -1,0 +1,3 @@
+package com.example.liverpooltest.model.others
+
+class MetaData
