+++ app/src/main/java/com/example/liverpooltest/model/others/RefinementGroup.kt
@@ -1,4 +1,4 @@
-package com.example.liverpooltest.model
+package com.example.liverpooltest.model.others
 
 data class RefinementGroup(
     val dimensionName: String,
