+++ app/src/main/java/com/example/liverpooltest/model/others/SortOption.kt
@@ -1,4 +1,4 @@
-package com.example.liverpooltest.model
+package com.example.liverpooltest.model.others
 
 data class SortOption(
     val label: String,
