+++ app/src/main/java/com/example/liverpooltest/ui/view/MainActivity.kt
@@ -1,25 +1,24 @@
 package com.example.liverpooltest.ui.view
 
 import android.os.Bundle
-import android.widget.Button
-import android.widget.EditText
-import android.widget.Spinner
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.liverpooltest.R
-import com.example.liverpooltest.model.ProductRepository
+import com.example.liverpooltest.databinding.ActivityMainBinding
+import com.example.liverpooltest.model.network.ProductRepository
 import com.example.liverpooltest.ui.viewmodel.ProductViewModel
 import com.example.liverpooltest.ui.viewmodel.ViewModelFactory
 
 class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
     private lateinit var viewModel: ProductViewModel
     private lateinit var adapter: ProductAdapter
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
         val repository = ProductRepository()
         viewModel = ViewModelProvider(
@@ -28,27 +27,22 @@ class MainActivity : AppCompatActivity() {
         )[ProductViewModel::class.java]
 
         adapter = ProductAdapter()
-        val recyclerView = findViewById<RecyclerView>(R.id.recyclerView)
-        recyclerView.layoutManager = LinearLayoutManager(this)
-        recyclerView.adapter = adapter
+        binding.recyclerView.layoutManager = LinearLayoutManager(this)
+        binding.recyclerView.adapter = adapter
 
-        val searchField = findViewById<EditText>(R.id.searchInput)
-        val button = findViewById<Button>(R.id.searchButton)
-        val sortSpinner = findViewById<Spinner>(R.id.sortSpinner)
-
-        button.setOnClickListener {
-            val query = searchField.text.toString()
-            val sortValue = getSortValue(sortSpinner.selectedItem.toString())
+        binding.searchButton.setOnClickListener {
+            val query = binding.searchInput.text.toString()
+            val sortValue = getSortValue(binding.sortSpinner.selectedItem.toString())
             viewModel.fetchProducts(query, sort = sortValue, reset = true)
         }
 
         viewModel.products.observe(this) { adapter.submitList(it.toList()) }
 
-        recyclerView.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+        binding.recyclerView.addOnScrollListener(object : RecyclerView.OnScrollListener() {
             override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
                 if (!recyclerView.canScrollVertically(1)) {
-                    val query = searchField.text.toString()
-                    val sortValue = getSortValue(sortSpinner.selectedItem.toString())
+                    val query = binding.searchInput.text.toString()
+                    val sortValue = getSortValue(binding.sortSpinner.selectedItem.toString())
                     viewModel.fetchProducts(query, sort = sortValue)
                 }
             }
