+++ app/src/main/java/com/example/liverpooltest/ui/viewmodel/ProductViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.liverpooltest.Product
-import com.example.liverpooltest.model.ProductRepository
+import com.example.liverpooltest.model.network.ProductRepository
 import kotlinx.coroutines.launch
 
 class ProductViewModel(private val repository: ProductRepository) : ViewModel() {
